requirements.txt
torch>=2.0.0
numpy>=1.24.0
Pillow>=9.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
scipy>=1.10.0
transformers>=4.30.0
huggingface-hub>=0.16.0
tqdm>=4.65.0 